+++ HadSvc/kmsginterface.h
@@ -1,0 +1,12 @@
+#pragma once
+class kMsgInterface
+{
+	inline
+		kMsgInterface() {}
+	inline
+		~kMsgInterface() {}
+
+
+	void kMsg_push();
+};
+
